optimising results page
